refactored model attributes as properties
so that attibute initialization does not happen in constructor. Indeed, branch2
needs a dataset that is generated by model.cropper. If the images have not been
cropped on model instanciation, self.branch2 will fail to initialize in train
mode. Hence, we delay all attribute model initializations until they are
called (rather than in the constructor). This allows instanciation of the model
on an brand new project